Add files via upload
This task is to move the rover from its landing site to one of the target sites for experiments and soil sampling. For an ideal rover that can cross every place, usually, the shortest geometrical path is defined as an optimal path.
A path is composed of a sequence of elementary moves. Each elementary move consists of moving the rover to one of its 8 neighbors. To find the solution you will use the following algorithms:
-Breadth-first search (BFS)
-Uniform-cost search (UCS)
-A* search(A*).